Use relative imports to fix use in TS projects. Closes #63
The `baseUrl` option in `tsconfig` should not be used in code bases that are meant to be consumed as libraries in other TypeScript-based projects, as the import paths in the emitted `d.ts` files will not be rewritten by the TypeScript compiler, and can not be resolved at build time. Thus the various types re-exported from the main `index.d.ts` module are exported as `any` and not really usable in practice.